requirements.txt
discord.py
requests
beautifulsoup4
python-dotenv
SQLAlchemy>=2.0
psycopg2-binary